runtime: consolidate message operations
