+++ internal/runtime/funcs/list_prepend.go
@@ -1,0 +1,48 @@
+package funcs
+
+import (
+	"context"
+
+	"github.com/nevalang/neva/internal/runtime"
+	"github.com/nevalang/neva/internal/runtime/messages"
+)
+
+type listPrepend struct{}
+
+//nolint:varnamelen // TODO(strict-lint phase 1): temporary suppression; remove after strict cleanup.
+func (p listPrepend) Create(io runtime.IO, _ messages.Msg) (func(ctx context.Context), error) {
+	dataIn, err := io.In.Single("data")
+	if err != nil {
+		//nolint:wrapcheck // TODO(strict-lint phase 1): temporary suppression; remove after strict cleanup.
+		return nil, err
+	}
+	lstIn, err := io.In.Single("lst")
+	if err != nil {
+		//nolint:wrapcheck // TODO(strict-lint phase 1): temporary suppression; remove after strict cleanup.
+		return nil, err
+	}
+
+	resOut, err := io.Out.Single("res")
+	if err != nil {
+		//nolint:wrapcheck // TODO(strict-lint phase 1): temporary suppression; remove after strict cleanup.
+		return nil, err
+	}
+
+	return func(ctx context.Context) {
+		for {
+			dataMsg, lstMsg, ok := receive2(ctx, dataIn, lstIn)
+			if !ok {
+				return
+			}
+
+			if !resOut.Send(
+				ctx,
+				messages.ListPrepend(lstMsg.List(), dataMsg.Msg),
+				dataMsg,
+				lstMsg,
+			) {
+				return
+			}
+		}
+	}, nil
+}
