+++ internal/runtime/funcs/stream_to_dict.go
@@ -41,8 +41,8 @@ func (streamToDict) Create(
 				continue
 			case isStreamData(dataMsg.Msg):
 				entryMsg := streamDataValue(dataMsg.Msg).Struct()
-				key := messages.StructGet(entryMsg, "key").Str()
-				valueMsg := messages.StructGet(entryMsg, "value")
+				key := messages.StructGetField(entryMsg, "key").Str()
+				valueMsg := messages.StructGetField(entryMsg, "value")
 
 				// Duplicate key policy: last message for the key wins.
 				dict[key] = valueMsg
