+++ internal/runtime/messages/bool.go
@@ -28,3 +28,12 @@ func equalBoolValue(left bool, right Msg) bool {
 	rightTyped, ok := right.(BoolMsg)
 	return ok && left == rightTyped.v
 }
+
+// BoolAnd returns the logical conjunction of left and right.
+func BoolAnd(left, right Msg) BoolMsg { return NewBoolMsg(left.Bool() && right.Bool()) }
+
+// BoolOr returns the logical disjunction of left and right.
+func BoolOr(left, right Msg) BoolMsg { return NewBoolMsg(left.Bool() || right.Bool()) }
+
+// BoolNot returns the logical negation of value.
+func BoolNot(value Msg) BoolMsg { return NewBoolMsg(!value.Bool()) }
